move creater mongoose model to entity
