+++ src/domain/commentsLikes-entity.ts
@@ -1,8 +1,10 @@
 import mongoose from 'mongoose'
-import { CommentsLikesTypeToDB } from '../models/likes-models'
+import {CommentsLikesTypeToDB} from '../models/likes-models'
 
 export const CommentsLikesSchema = new mongoose.Schema<CommentsLikesTypeToDB>({
     userId: { type: String, required: true },
     commentId: { type: String, required: true },
     myStatus: { type: String, required: true }
 })
+
+export const CommentsLikesModel = mongoose.model('comments-likes', CommentsLikesSchema)
