+++ src/domain/rateLimiter-entity.ts
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose'
-import { RateLimiterTypeToDB } from '../models/auth-models'
+import {RateLimiterTypeToDB} from '../models/auth-models'
 
 export const RateLimiterSchema = new mongoose.Schema<RateLimiterTypeToDB>({
     ip: { type: String, required: true },
@@ -7,3 +7,5 @@ export const RateLimiterSchema = new mongoose.Schema<RateLimiterTypeToDB>({
     date: { type: Number, required: true },
     count: { type: Number, required: true }
 })
+
+export const RateLimiterModel = mongoose.model('limiters', RateLimiterSchema)
