+++ src/query-reositories/blogs-query-repository.ts
@@ -1,10 +1,10 @@
 import { BlogsTypeOutput } from '../models/blogs-models'
-import { BlogsModel } from '../repositories/db/db'
 import { QueryBlogs } from '../models/query-models'
 import { PaginatedType } from '../models/main-models'
 import { getPaginatedType, makeDirectionToNumber } from './helper'
 import { injectable } from 'inversify'
 import { ObjectId } from 'mongodb'
+import {BlogsModel} from "../domain/blogs-entity";
 
 @injectable()
 export class BlogsQueryRepository {
