+++ src/query-reositories/comments-query-repository.ts
@@ -1,12 +1,13 @@
 import { QueryComments } from '../models/query-models'
 import { PaginatedType } from '../models/main-models'
 import { CommentsTypeOutput } from '../models/comments-models'
-import { CommentsModel, PostsModel } from '../repositories/db/db'
 import { getPaginatedType, makeDirectionToNumber } from './helper'
 import { ObjectId } from 'mongodb'
 import { CommentsLikesRepository } from '../repositories/comments-likes-repository'
 import { inject, injectable } from 'inversify'
 import { CommentsLikesQueryRepository } from './comments-likes-query-repository'
+import {PostsModel} from "../domain/posts-entity";
+import {CommentsModel} from "../domain/comments-entity";
 
 @injectable()
 export class CommentsQueryRepository {
