+++ src/repositories/comments-likes-repository.ts
@@ -1,7 +1,7 @@
-import { CommentsLikesModel } from './db/db'
-import { CommentsLikesId, CommentsLikesTypeToDB, StatusLikeType } from '../models/likes-models'
-import { ObjectId } from 'mongodb'
-import { injectable } from 'inversify'
+import {CommentsLikesId, CommentsLikesTypeToDB, StatusLikeType} from '../models/likes-models'
+import {ObjectId} from 'mongodb'
+import {injectable} from 'inversify'
+import {CommentsLikesModel} from "../domain/commentsLikes-entity";
 
 @injectable()
 export class CommentsLikesRepository {
