+++ src/repositories/test-repository.ts
@@ -1,14 +1,12 @@
-import {
-    BlogsModel,
-    CommentsLikesModel,
-    CommentsModel,
-    PostsLikesModel,
-    PostsModel,
-    RateLimiterModel,
-    SessionsModel,
-    UsersModel
-} from './db/db'
 import { injectable } from 'inversify'
+import {BlogsModel} from "../domain/blogs-entity";
+import {PostsModel} from "../domain/posts-entity";
+import {UsersModel} from "../domain/users-entity";
+import {CommentsModel} from "../domain/comments-entity";
+import {SessionsModel} from "../domain/sessions-entity";
+import {RateLimiterModel} from "../domain/rateLimiter-entity";
+import {CommentsLikesModel} from "../domain/commentsLikes-entity";
+import {PostsLikesModel} from "../domain/postsLikes-entity";
 
 @injectable()
 export class TestRepository {
