+++ src/services/users-service.ts
@@ -2,7 +2,7 @@ import { inject, injectable } from 'inversify'
 import { UsersRepository } from '../repositories/users-repository'
 import { UserId } from '../models/users-models'
 import { UsersQueryRepository } from '../query-reositories/users-query-repository'
-import { UsersModel } from '../repositories/db/db'
+import {UsersModel} from "../domain/users-entity";
 
 @injectable()
 export class UsersService {
